Rename git files
