+++ src/commands/react-native/upload.ts
@@ -5,7 +5,7 @@ import asyncPool from 'tiny-async-pool'
 import {ApiKeyValidator, newApiKeyValidator} from '../../helpers/apikey'
 import {getBaseSourcemapIntakeUrl} from '../../helpers/base-intake-url'
 import {InvalidConfigurationError} from '../../helpers/errors'
-import {getRepositoryData, newSimpleGit, RepositoryData} from '../../helpers/git-sourcemaps-repository-data'
+import {getRepositoryData, newSimpleGit, RepositoryData} from '../../helpers/git/format-git-sourcemaps-data'
 import {RequestBuilder} from '../../helpers/interfaces'
 import {getMetricsLogger, MetricsLogger} from '../../helpers/metrics'
 import {upload, UploadStatus} from '../../helpers/upload'
